Preparatory refactors
