+++ apps/patient/src/components/PharmacyInfo.tsx
@@ -22,7 +22,7 @@ import { useMemo, useState } from 'react';
 import { IoChevronDownOutline, IoChevronUpOutline } from 'react-icons/io5';
 import { formatAddress, formatPrice, titleCase } from '../utils/formatters';
 import { getFulfillmentTrackingLink } from '../utils/fulfillmentsHelpers';
-import { AmazonOffer, BrandedOptionOverrides } from './BrandedOptions';
+import { AmazonOffer, BrandedOptionOverrides } from './pharmacy-card-list/BrandedOptions';
 
 dayjs.extend(customParseFormat);
