+++ apps/patient/src/components/index.tsx
@@ -1,5 +1,3 @@
-export * from './BrandedOptions';
-export * from './BrandedPharmacyCard';
 export * from './DemoCtaModal';
 export * from './FixedFooter';
 export * from './LocationModal';
