+++ apps/patient/src/components/pharmacy-card-list/BrandedPharmacyCard.tsx
@@ -1,15 +1,15 @@
 import { Card, CardBody } from '@chakra-ui/react';
 
-import capsuleLogo from '../assets/capsule_logo_small_circle.png';
-import amazonPharmacyLogo from '../assets/amazon_pharmacy_logo_small_circle.png';
-import altoLogo from '../assets/alto_logo.svg';
-import costcoLogo from '../assets/costco_logo_small.png';
-import costPlusLogo from '../assets/costplus_logo_small_circle.png';
-import walmartLogo from '../assets/walmart_logo_small_circle.png';
-import novocareLogo from '../assets/novo_circle.png';
+import capsuleLogo from '../../assets/capsule_logo_small_circle.png';
+import amazonPharmacyLogo from '../../assets/amazon_pharmacy_logo_small_circle.png';
+import altoLogo from '../../assets/alto_logo.svg';
+import costcoLogo from '../../assets/costco_logo_small.png';
+import costPlusLogo from '../../assets/costplus_logo_small_circle.png';
+import walmartLogo from '../../assets/walmart_logo_small_circle.png';
+import novocareLogo from '../../assets/novo_circle.png';
 
-import capsulePharmacyIdLookup from '../data/capsulePharmacyIds.json';
-import { PharmacyInfo } from './PharmacyInfo';
+import capsulePharmacyIdLookup from '../../data/capsulePharmacyIds.json';
+import { PharmacyInfo } from '../PharmacyInfo';
 import { BrandedOptionOverrides } from './BrandedOptions';
 
 interface Props {
