+++ apps/patient/src/components/pharmacy-card-list/index.ts
@@ -1,1 +1,3 @@
 export * from './PickupPharmacyCardList';
+export * from './BrandedOptions';
+export * from './BrandedPharmacyCard';
